% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-LondonHPI}
\alias{LondonHPI}
\alias{data-LondonHPI}
\title{London Housing Price Index, 1995--2015}
\format{A univariate time series with 252 observations; a \code{ts} object}
\description{
Contains 252 monthly index values for the years 1995--2015 of the UK Housing
price index for London.
}
\details{
The data was downloaded from the 'UK House Price Index' website.

To retrieve the data: go to the `customise your search' part of the
`search the UK house price index' form. There select the `English region'
London, the period from 01-1995 to 12-2015, and obtained the `average price'
for `all property types'.
}
\examples{
## compute monthly changes of average prices
L <- length(LondonHPI)
LondonHPI_change <- LondonHPI[2:L]/LondonHPI[1:(L-1)] - 1
## demean
LondonHPI_mean <- mean(LondonHPI_change)
LondonHPI_adj <- ts(LondonHPI_change - LondonHPI_mean, f = 12, start=c(1995,2))
op <- par(mfcol=c(1,2))
plot(LondonHPI, main = "", ylab="average prices", xlab="year")
plot(LondonHPI_adj, main = "", ylab="monthly changes", xlab="year")
par(op)
}
\references{
UK House Price Index
\url{http://landregistry.data.gov.uk/app/ukhpi}
}
\keyword{data}

