% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tvARMAcpp}
\alias{tvARMAcpp}
\title{Workhorse function for tvARMA time series generation}
\arguments{
\item{z}{a ...}

\item{x_int}{a ...}

\item{A}{...}

\item{B}{a ...}

\item{Sigma}{a ...}
}
\value{
Returns a ...
}
\description{
More explanation!
}
